requirements.txt
pydriller
pybars4
importlib_resources
wrapt